refactor: remove redundant study model
