+++ src/entities/preferences/index.ts
@@ -2,7 +2,6 @@ export { usePreferences } from "./model/hooks";
 export { studyPreferencesLimits } from "./model/rules";
 export type {
   Preferences,
-  SwipeAction,
   SwipeDirection,
 } from "./model/types";
 export { setDarkMode, toggleShowHeader, toggleShowSwipeButtonList, updatePreferences } from "./model/store";
