+++ src/features/study/index.ts
@@ -1,2 +1,2 @@
-export { DeckSwiperView } from "./components/DeckSwiperView";
+export { DeckSwiperView } from "./ui/DeckSwiperView";
 export { useStudy, type StudyState } from "./hooks/useStudy";
