+++ src/features/study/ui/DeckSwiperView.tsx
