+++ src/features/study/ui/SwipeButtonList.spec.tsx
