+++ src/features/study/ui/SwipeButtonList.tsx
