+++ test/factories.ts
@@ -8,11 +8,12 @@ import type { RemoteCard } from "@/entities/card";
 import type { LocalCard } from "@/entities/card/model/types";
 import type { RemoteDeck } from "@/entities/deck";
 import type { LocalDeck } from "@/entities/deck/model/types";
-import type { Preferences, SwipeAction } from "@/entities/preferences";
+import type { Preferences } from "@/entities/preferences";
 
 type AppearancePreferences = Preferences["appearance"];
 type StudyPreferences = Preferences["study"];
 type ControlPreferences = Preferences["controls"];
+type SwipeAction = ControlPreferences["cardSwipeUp"];
 
 /**
  * Builds a complete test deck with predictable defaults and optional field overrides.
